Make the page usable at limit

+++ php/adminBookRemove.php
@@ -2,15 +2,20 @@
 include "dbConnection.php";
 session_start();
 
-$bookId;
+
 
 if(!empty($_SESSION['adminName'])){
-    $remove_query="DELETE FROM books 
-    WHERE ISBN = '$bookId'";
-    if(mysqli_query($connection,$remove_query)){
-        echo '<script>alert("Remove sucessfully!")</script>';
-    }else{
-        echo '<script>alert("Process failed")</script>';
+
+    if(isset($_POST['removeBook'])){
+        $bookId=$_POST['ISBN'];
+        $remove_query="DELETE FROM books 
+        WHERE ISBN = '$bookId'";
+        if(mysqli_query($connection,$remove_query)){
+            echo '<script>alert("Remove sucessfully!")</script>';
+        }else{
+            echo '<script>alert("Process failed")</script>';
+        }
     }
+
 }
 ?>
